Make inner class TotalPowerAndTotalRolls a top level class (#7841)
