+++ game-core/src/main/java/games/strategy/triplea/ai/pro/util/ProSortMoveOptionsUtils.java
@@ -9,10 +9,10 @@ import games.strategy.engine.data.UnitType;
 import games.strategy.triplea.ai.pro.ProData;
 import games.strategy.triplea.ai.pro.data.ProTerritory;
 import games.strategy.triplea.attachments.UnitAttachment;
-import games.strategy.triplea.delegate.DiceRoll;
 import games.strategy.triplea.delegate.Matches;
 import games.strategy.triplea.delegate.TerritoryEffectHelper;
 import games.strategy.triplea.delegate.battle.UnitBattleComparator;
+import games.strategy.triplea.delegate.dice.TotalPowerAndTotalRolls;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.LinkedHashMap;
@@ -214,8 +214,8 @@ public final class ProSortMoveOptionsUtils {
         }
         powerDifference +=
             (includeUnit ? 1 : -1)
-                * DiceRoll.getTotalPower(
-                    DiceRoll.getUnitPowerAndRollsForNormalBattles(
+                * TotalPowerAndTotalRolls.getTotalPower(
+                    TotalPowerAndTotalRolls.getUnitPowerAndRollsForNormalBattles(
                         sortedUnits, defendingUnits, sortedUnits, false, data, t, effects),
                     data);
       }
